+++ server/api/address.js
@@ -1,4 +1,5 @@
 const Address = require('../db/models/address')
+const router = require('express').Router()
 
 router.put('/:id', async (req, res, next) => {
   const id = req.params.id
