+++ server/api/index.js
@@ -1,6 +1,6 @@
 const router = require('express').Router()
 
-router.use('/orders', require('./order'))
+router.use('/orders', require('./orders'))
 router.use('/users', require('./users'))
 router.use('/products', require('./products'))
 router.use('/address', require('./address'))
